Increase Themed Icon Size
In this CL we found that the current size of the icons does not correctly account for the size of the key shadow, which makes them slightly larger than the themed icons. We choose instead to increase the size of the themed icons and cut the shadow off slightly to be able to more easily fix the issue.

Bug: b/298203449
Test: TBD
Flag: None
Change-Id: Ia9ae7a1c20be24c639a1dc3542984ebd12c2523a