Support flow value tracing
Also, change format of flow trace strings to use "#" instead of ":".

":" has special meaning in the Perfetto UI making it more difficult to
search for occurrences.

Fix tests, and throw errors for invalid trace sections closer to the
expect() function so that its easier to locate the line of code on which
the test failed.

Test: atest tracinglib-robo-test
Test: setprop persist.debug.coroutine_tracing.flow_values true
Flag: com.android.systemui.coroutine_tracing
Bug: 289353932
Change-Id: Ib3d14e81f2f540757d76f04bec1c5186c6d88030