Allow the type of semantic to be specified when creating the effect.
This is in line with what we do in the Toggle, but we have not yet
upgraded the `MagneticDetach` effect. For now, we introduce an
factory fun / object indirection, so that we can specify the effect
semantic type via `MagneticDetachEffect`, while all existing references
still work.

Test: Existing Unit tests
Bug: 419520966
Bug: 448605986
Flag: EXEMPT PURE_REFACTOR
Change-Id: I392d1c4cfe4509e50f5e62fd48fab3f7e8df021d