MM: Update VerticalTactileSurfaceRevealTest for programmatic animations
This commit refactors `VerticalTactileSurfaceRevealModifierTest` to test
programmatically triggered animations (like `animateToScene` or
`showOverlay`) instead of user-driven gestures.

As part of this test update the `isTransitioning` feature captured for
the golden test is corrected to use a boolean type instead of an int.

Test: atest VerticalTactileSurfaceRevealModifierTest
Bug: 392535471
Flag: com.android.systemui.qs_tiles_reveal_animation
Change-Id: I1cd4bf42b3b0974abb3b9ad9ff98c56626e55267