tracinglib: adjust invokeSuspend naming pattern
Include functions that start with "invokeSuspend" instead of only those
that are exactly equal to "invokeSuspend".

It's common for suspend calls to be compiled into a form such as:
"invokeSuspend$lambda$2". These should be included as well when trying
to infer a name from the stack trace.

Bug: 289353932
Flag: EXEMPT DEBUG
Test: atest tracinglib-robo-test
Change-Id: I480e6c49be54d2858d723da1d8c22e70d05f8aa3